+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## (5.1.0 IN PROGRESS)
 
+* Configure Plugin to find privileged contacts. Refs UIPFCONT-37.
+
 ## [5.0.0](https://github.com/folio-org/ui-plugin-find-contact/tree/v5.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-contact/compare/v4.0.0...v5.0.0)